changed to int

+++ model/bhac/parameters.h
@@ -113,7 +113,7 @@ double M_UNIT;
 //////////////////////
 
 double CAM_FREQ;
-double TIME_INIT;
+int TIME_INIT;
 double INCLINATION;
 
 // SED parameters
